% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_cont_disc.r
\name{random.sample}
\alias{random.sample}
\title{Spatially random sample}
\usage{
random.sample(
  bounding_geom = NULL,
  key = NULL,
  value = NULL,
  data_return = c("osm_polygons", "osm_points", "osm_multipolygons", "multilines",
    "lines"),
  boundary = 0,
  buff_dist = 0,
  buff_epsg = 4326,
  join_type = "within",
  dis_or_cont,
  sample_size,
  plotit = TRUE,
  plotit_leaflet = TRUE
)
}
\arguments{
\item{bounding_geom}{a \code{sf} or \code{sp} object (with \eqn{N \geq \code{size}}) where each line corresponds to one spatial location. It should contain values of 2D coordinates, data and, optionally, covariate(s) value(s) at the locations. This argument must be provided when sampling from a \code{"discrete"} set of points, see \code{'type'} below for details.}

\item{key}{aas}

\item{value}{aas}

\item{data_return}{aas}

\item{boundary}{aas}

\item{buff_dist}{aas}

\item{buff_epsg}{aas}

\item{join_type}{aas}

\item{dis_or_cont}{random sampling, a choice of either \code{"discrete"}, from a set of \eqn{N} potential sampling points or \code{"continuum"} from independent, compeletely random points.}

\item{sample_size}{a non-negative integer giving the total number of locations to be sampled.}

\item{plotit}{'logical' specifying if graphical output is required. Default is \code{plotit = TRUE}.}

\item{plotit_leaflet}{'logical' specifying if leaflet (html) graphical output is required. This is prioritised over plotit if both are selected. Default is \code{plotit_leaflet = TRUE}.}
}
\value{
a \code{df} object of dimension \eqn{n} by \code{4} containing the final sampled osm_ids, centroid locations and whether the instance is in the selected sample (0/1), if sampling from a \code{"discrete"} set of points. A \code{df} object of dimension \eqn{n} by \code{3} containing the serial id and centroid locations for all sample instances,if sampling from a \code{"continuum"}.
}
\description{
This function draws a spatially random sample from a either (1) a discrete set of OSM features defined in the function parameters or (2) a continuous surface defined by a user definted geographical region.
}
\examples{
set.seed(15892)
xy.sample <-

set.seed(15892)
xy.sample <-

set.seed(15892)
xy.sample


}
\references{
Rowlingson, B. and Diggle, P. 1993 Splancs: spatial point pattern analysis code in S-Plus. Computers and Geosciences, 19, 627-655
}
\author{
Henry J. Crosby \email{henry.crosby@warwick.ac.uk}

Godwin Yeboah \email{godwin.yeboah@warwick.ac.uk}

J. Porto De Albuquerque \email{J.Porto@warwick.ac.uk}
}
